Update ru_ru-ru.csv
Edited using LLM, checked and brought to the colloquial style of the language manually. Not tested in the game. Some phrases are adapted to take into account the peculiarities of the Russian language. Typos and punctuation errors are possible. The basic style of the text, jokes and narration is preserved.